Narrow down long list of trading symbols to be useable
This script reads from the "pairs_on_exchanges.py" dictionary and returns a new dictionary of exchanges that ONLY includes the symbols that are common amongst other exchanges.
The "pairs_on_exchanges.py" file will be updated to reflect this.

It also outputs the number of symbols trading on each exchange.