Made 'get_trading_symbols.py' a script that gets the most up-to-date symbols that are compatible to monitor and trade for arbitrage
It uses CCXT to automatically select pairs that are shared between exchanges (to avoid wasted monitoring) and automatically excludes any fiat pairings.
Gives a dictionary that has exchanges as keys and their respective pairs as values.

In future, will get rid of some of the print statements as this will be imported into larger programs and the information won't be important for the user.