Completed initialise_exchanges script
When using in the main program, simply import initialise_exchanges as init_ex and then declare the dictionary exchanges = init_ex.exchanges.

This module is based off 'designate_workload.py' with 2 extra lines at the end making a call to declare the dictionary. The old script will be removed in this commit.