Fixed numerous bugs
In 'orderbook_hdf.py', a function called the symbol and exchange object in the wrong place. As well, ccxt has some inherent flaws, most notably the precision is missing in some cases. For this, a quick fix of letting precision for both price and amount = 8 digits was done. This will have to be solved more rigorously when actually placing live trades. If ccxt does not upgrade their precision attributes by the time the live trading bot is made, there will have to be conditional statements with precisions attained manually for some exchanges.

Another fix in 'orderbook_hdf.py' is the removal of symbols from the exchanges dictionary if the API call to that symbol isn't supported. This is because for some coins, calls to the orderbook have become deprecated and ccxt developers have not removed them yet. This is something the program simply has to accept and deal with accordingly.

In 'get_trading_symbols.py', the exchange Cryptopia has been removed.