Automated SSH communication
A new script, 'file_transfer.py',  is used (named 'file_transfer_public.py' here with credentials removed). This is essentially a warpper used for automation of server tasks, as there are many servers that will be run simultaneously.

With this script, the following can be done seamlessly:
- Dependencies installed and the necessary 'orderbook' folder created
- All files copied to server's directory
- ServerNumber allocated
- Orderbook collection program set to run in background with nohup command (could not use screen)
- All stdout text logged.

This script is run locally with Python. What's more, when retrieving the data in future, there is already a function defined to handle that, 'getFilesFromServer()' so the process will be quick and automated.