Fixed timestamp = None bug
CCXT removed timestamps from most symbols in the fetch_l2_orderbook() method, and because of a conditional statement in the code, when orderbook['timestamp'] == None, the symbol would be skipped. To deal with this, the condition was changed to if not orderbook['asks']. The timestamp was then gotten with the time module and included in the dataframe.

Also made program enter data to folders named 'orderbook0', 'orderbook1', etc., depending on the server it's running from.